+++ EconomicsAPI/Utils/SpawnProjectile.cs
@@ -34,8 +34,8 @@ public class SpawnProjectile
         }
         Projectile projectile = Main.projectile[num];
         projectile.SetDefaults(Type);
-        projectile.position.X = X - (float)projectile.width * 0.5f;
-        projectile.position.Y = Y - (float)projectile.height * 0.5f;
+        projectile.position.X = X - projectile.width * 0.5f;
+        projectile.position.Y = Y - projectile.height * 0.5f;
         projectile.owner = Owner;
         projectile.velocity.X = SpeedX;
         projectile.velocity.Y = SpeedY;
@@ -68,17 +68,17 @@ public class SpawnProjectile
             switch (Type)
             {
                 case 206:
-                    projectile.ai[0] = (float)Main.rand.Next(-100, 101) * 0.0005f;
-                    projectile.ai[1] = (float)Main.rand.Next(-100, 101) * 0.0005f;
+                    projectile.ai[0] = Main.rand.Next(-100, 101) * 0.0005f;
+                    projectile.ai[1] = Main.rand.Next(-100, 101) * 0.0005f;
                     break;
                 case 335:
                     projectile.ai[1] = Main.rand.Next(4);
                     break;
                 case 358:
-                    projectile.ai[1] = (float)Main.rand.Next(10, 31) * 0.1f;
+                    projectile.ai[1] = Main.rand.Next(10, 31) * 0.1f;
                     break;
                 case 406:
-                    projectile.ai[1] = (float)Main.rand.Next(10, 21) * 0.1f;
+                    projectile.ai[1] = Main.rand.Next(10, 21) * 0.1f;
                     break;
                 default:
                     projectile.ai[0] = ai0;
